#11 - 리뷰 저장 api 구현
form 으로 전달되는 것 외에
작성자 이름, 작성 시간을 저장하고
리뷰 작성 후 파홀,숙소,호스트 평점 update 하도록 구현

+++ src/main/java/com/example/farmingholiday/domain/House.java
@@ -73,5 +73,5 @@ public class House {
     private double rate;
 
     @Column
-    private double reviewCount;
+    private Long reviewCount;
 }
